requirements.txt
hydra-core<1.1
joblib
nnmnkwii >= 0.0.23
nnsvs
numpy >= 1.20
omegaconf
scipy
torch
tqdm
utaupy >= 1.10
pyyaml
